+++ guillotina/json/deserialize_content.py
@@ -73,21 +73,22 @@ class DeserializeFromJson(object):
 
         factory = get_cached_factory(self.context.type_name)
         main_schema = factory.schema
-        await self.set_schema(
+        changed = await self.set_schema(
             main_schema, self.context, data, errors, validate_all, False)
 
         if errors and not ignore_errors:
             raise DeserializationError(errors)
 
-        self.context._p_register()
+        if changed:
+            self.context._p_register()
 
         return self.context
 
     async def set_schema(
             self, schema, obj, data, errors,
-            validate_all=False, behavior=False):
+            validate_all=False, behavior=False) -> bool:
         write_permissions = merged_tagged_value_dict(schema, write_permission.key)
-
+        changed = False
         for name, field in get_fields(schema).items():
             if name in RESERVED_ATTRS:
                 continue
@@ -128,6 +129,7 @@ class DeserializeFromJson(object):
                     # record object changes for potential future conflict resolution
                     try:
                         await apply_coroutine(field.set, obj, value)
+                        changed = True
                     except ValidationError as e:
                         errors.append({
                             'message': e.doc(), 'field': name, 'error': e})
@@ -144,6 +146,7 @@ class DeserializeFromJson(object):
                                 'Error setting data on field, falling back to setattr',
                                 exc_info=True)
                             setattr(obj, name, value)
+                            changed = True
                         else:
                             logger.warning(
                                 'Error setting data on field', exc_info=True)
@@ -169,6 +172,7 @@ class DeserializeFromJson(object):
                         'field': error[0],
                         'error': error
                     })
+        return changed
 
     async def get_value(self, field, obj, value):
         if value is None: